nixl/core: New infra folder to separate helpers
Moved the various descriptors, memory secitons, metadata to their own
infra folder. The agent and plugin manager are in core. Performed this
split because all plugins depend on the infra part but not on the agent.
But the plugin manager depends on static plugins. We had this split
before already but now its more cleaner.

Adjusted the meson builds for everything else.

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>