Implement multi-threading support
Current NIXL implementation sidesteps any considerations of executing in
multi-threaded environment by leaving it up to the user to ensure that
access to nixlAgent instance is synchronized. Extend this by adding a
strict synchronization mode while also preserving existing mode but making
it explicit.

To continue using NIXL with exactly the same performance set configuration
variable disable_mt=true which statically removes all locking at build
time.

When NIXL is compiled with multi-threading enabled still allow dynamic
configurations via nixlAgentConfig->syncMode field:

- When set to NIXL_SYNC_NONE no locking is performed and behavior is
similar as before this change with the small difference that dynamic check
is performed every time the lock would be taken with strict model.

- When set to NIXL_SYNC_STRICT all nixlAgent API methods are fully
synchronized with internal exclusive lock.

In order to allow dynamic configuration at runtime the lock implemented as
a thin BasicLockable-compatible wrapper around std::mutex that accepts
sync_mode parameter to its constructor and skip locking/unlocking the mutex
when synchronization mode is not strict. To allow static configuration at
build time the lock is compiled out and the macro that is used to create a
lock guard for the lock is set to no-op.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>