Make nixlPluginManager initialization thread-safe
Currently the initialization of nixlPluginManager class is split between
constructor and registerBuiltinPlugins() member function that is called
only once in getInstance(). This doesn't seem to be caused by any
necessity, so just call the init function from ctor and remove all the
machinery required to call it only once in a multi-threaded environment.

The call to ctor of static function-local instance of nixlPluginManager is
already thread safe in C++11 and later (Meyers singleton).

While we are at it, also make registerStaticPlugin() method private since
it is only called from registerStaticPlugin which is now just a constructor
helper and remove the locking from it.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>