listener: Separate etcd and listener thread in comm work
Etcd is independent of the listener thread so lets use the right flag to
start listener thread if required. Comm work thread will handle the etcd
requests from agent.

For invalidate keys, a .invalid key is added instead. When the invalid
key is fetched, it is deleted.

Use the new NIXL logging functions for etcd output.

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>