Add 'const' modifier to agent data path methods
For NIXL to allow executing a subset of data path APIs in parallel the
agent needs to become aware of the concurrency and ensure that it doesn't
unsafely modify shared state in implementations of those methods. Previous
commit in this series marked engine methods that are called by
reateXferReq, postXferReq, getXferStatus, releaseXferReq, genNotif methods
const, now also add 'const' modifier to getXferStatus, releaseXferReq and
genNotif since the other two are already constant methods.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>