Extend UCX backend to support multiple workers
Configure number of workers by setting new agent custom parameter
'num_workers'.

Spawn multiple workers in UCX backend and establish connection on all of
them but decide which instance to use for data-path operations based on
caller thread id hash to allow equal spread of tasks over multiple workers
in a way that is transparent to the user.

For the purposes of generating metadata the first worker is always chosen.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>