Add NIXL Telemetry infrastructure
This commit introduces infrastructure and basic agent telemetry collection.
Telemetry data is written to a shared memory and can be read
and proccessd by other applications. A simple reader is added.
This commit adds a cyclic buffer and a Thread executer that
can be used across NIXL for other use cases.

Signed-off-by: Rafi Wiener <rwiener@nvidia.com>