Fix nixlbench: batch size greater than 1, proper file offsets
The previous change done here, should have retained the creation of remote_iovs
from local_iovs so that the amount of bytes to transfer in both lists is
identical.

We just need to make sure the remote IOV has the FD of the file, a file
offset and proper size. Before the PR, all batches wrote to file offset 0, which
was a bug, because the file size was calculated to allow for all buffers having
their own offset.

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>