[CI] verification image build job fixup
Images built with this job started to fail when run without any
arguments which was suppose to use nvidia_entrypoint script to just run
bash shell.

The error we got was:
/opt/nvidia/nvidia_entrypoint.sh: line 55: exec: --: invalid option
exec: usage: exec [-cl] [-a name] [command [argument ...]] [redirection ...]

It seems like the image got corrupet by adding version file to the built image
because it was done wrong.

This PR fixes the issue by changing the way we start the container from
docker create to docker run.

Signed-off-by: Daniel Pressler <danielpr@nvidia.com>