Optimize libfabric request tracking with atomic operations
Since postXfer is in the fast path, we can calculate the total number of requests in
a single pass as we submit each request. By initially assigning the total number of request
to be a maximum number of requests possible we can avoide the race condition
when we check fi all the operations are completed.