benchmark/nixlbench: Fix POSIX backend API param
The nixl_worker was mistakenly adding a boolean value to the POSIX API
parameter string map. As a result, when the POSIX backend tried to
retrieve the queue type, it failed to compare the boolean with a string
and always fell back to the AIO backend.

This patch fixes the issue.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>