nixlbench: Fix gusli parameters (#910) (#929)
* nixlbench: Fix gusli parameters

Fix the hardcoded dev id. Fix consistency check for GUSLI R/W for file.
The GUSLI file parser logic is moved to utils since utils will also use
it when checking for consistency. For consistency check for READ
operation, we pre-create the file and write data to it. Add a workaround
for GUSLI resetting DRAM when registering memory.

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>
Co-authored-by: Adit Ranadive <aranadive@nvidia.com>