benchmark/nixlbench: --config_file param supported
This patch introduces config file support to nixlbench.

The name of a config file can be specified using the --config_file
command-line parameter. The config file is in INI format.

Each existing command-line parameter can also be placed in the
"[global]" section of the configuration file, so the following
invocations:

  nixlbench --etcd_endpoints http://localhost:2379 --backend POSIX
    --filepath /mnt/test --posix_api_type AIO

and

  nixlbench --config_file /tmp/nixlbench.config

where /tmp/nixlbench.config contains:

[global]
etcd_endpoints=http://localhost:2379
backend=POSIX
filepath=/mnt/test
posix_api_type=AIO

are identical.

If a parameter exists in the config file and is also explicitly
specified on the command line, the latter takes precedence.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>