plugins/posix: minimized IO queue abstraction
This patch set minimizes the IO queue abstraction, making it as low-
level as possible.

It now only wraps the lowest-level IO-related API. It works with a set
of IOs (writes or reads) while leaving the Xfer level logic to the upper
layer (posix_backend).

This patch set allows the posix_backend to implement the common logic,
while only offloading the details of specific file access API to the IO
queue abstraction.

This patch set prepares the code for the POSIX plugin threading model
implementation. With it, the threading will be implemented once,
in the posix_backend.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>