plugins/posix: AIO resubmission fixed
The Linux AIO plugin wasn’t capable of re-submission; it required IO
to be re-prepared prior to re-submission.

It also used to return NIXL_IN_PROG for the next submit(), and
NIXL_SUCCESS for checkCompleted(), making the upper layer think that
the re-submission succeeded. This, in turn, caused incorrect
throughput-related calculations.

This patch fixes the bug.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>