CI: fix for GPU test random tcp port
The randon tcp port logic depends on executer number environment from
jenkins, since we are running inside docker container we do not have
this variable by default.
This is causing randon failures due to port collision.
This commit makes sure we provide the container the EXECUTOR_NUMBER
variable when we create the container.

Signed-off-by: Daniel Pressler <danielpr@nvidia.com>