Replace std::unordered_map with std::vector in nixlDlistH
Replace the unordered_map with a vector of pairs to improve performance
and reduce memory overhead for the typical case of 1-5 backends.

Since the number of backends is small (usually one), the overhead of
hashing and chasing pointers through a tree isn't worth it. Usually
it takes 20-30 elements before a tree beats a simple vector.

Signed-off-by: Ben Walker <ben@nvidia.com>