Use flexible array member for exact-size allocation
Replace the fixed-size std::array with a zero size array
that allocates exactly the space needed for the number of backends.

Overloaded operator new takes num_backends and allocates base + array

This makes the inline array added in the previous commit no longer have
a maximum size, at the cost of extra complexity when calling new().

Signed-off-by: Ben Walker <ben@nvidia.com>