benchmark/nixlbench: switched to cxxopts
This patch replaces gflags-based parameter parsing with cxxopts-based
parsing, and removes the gflags dependency.

This change will enable us to support config file-based parameters in
upcoming patches.

Unfortunately, gflags is not flexible enough for our needs. For example,
it does not indicate whether a parameter was explicitly provided by the
user or if it is using a default. This functionality is required for
correct parameter specification priority: first, a value provided by the
user on the command line; then a value from the config file; and finally,
the default value.

Additionally, gflags does not support dynamically formed arguments,
which is needed to handle a list of tasks as discussed. For this use
case, the config file would look like:

[general]
...
num_tasks=2

[task0]
local=...
remote=...
type=write
...

[task1]
local=...
remote=...
type=read
...

This translates to command line parameters like:

--num_tasks=2 --task0.local=... ... --task1.local=...

Since gflags is a compile-time mechanism, it cannot support this level
of dynamic argument handling.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>