nixl_ep: Metadata exchange via torch TCPStore
Most inference engines use their own communication
groups and prefer not to launch an additional ETCD
server solely for nixl_ep's connection establishment
flow.

This commit adds support for exchanging metadata
using an external torch.distributed.TCPStore
instance (as used by vLLM).

Note that this is likely an intermediate solution.
Further discussion is needed for the broader NIXL
approach, but the nixl_ep API changes are not
expected to change.

Signed-off-by: Itay Alroy <ialroy@nvidia.com>